LayoutTests/fast/profiler/profiling-from-a-nested-location-but-stop-profiling-outside-the-nesting.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.setJavaScriptProfilingEnabled(true);
}

function startTest()
{
    functionWichStartsTheProfiler();
    endTest();
}

function functionWichStartsTheProfiler()
{
    console.profile("Profiling From A Nested Location But Stop Profiling Outside The Nesting");

    for (var i = 0; i < 10000000; i++)
        var b = i + 2;
}

</script>
</head>

<body onload="startTest()">
This page's JavaScript starts profiling from within a nested location but stops the profiler
from a different level of nesting.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be a call to functionWichStartsAndStopsTheProfiler().
<div id="output"></div>
</body>
</html>
